requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
idna==3.10
numpy==2.2.3
opencv-python==4.11.0.86
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
python-dotenv==1.1.0
psycopg2-binary==2.9.10
colorlog=6.9.0